app/requirements.txt
albumentations
Flask
Pillow
timm
numpy
gunicorn~=20.1.0
